Don't drop when snippets are nested

+++ denops/denippet/main.ts
@@ -157,7 +157,8 @@ export function main(denops: Denops): void {
       session.unguard();
       if (
         await g.get(denops, "denippet_drop_on_zero", false) &&
-        session.snippet?.currentNode().tabstop === 0
+        session.snippet?.currentNode().tabstop === 0 &&
+        session.snippet?.outer == null
       ) {
         await session.drop();
       }
